refactor(animations): Treeshakable animation providers.
Use `providedIn:'root'` when possible to improve treeshaking for the animation module
