+++ packages/animations/browser/BUILD.bazel
@@ -14,6 +14,7 @@ ng_module(
     ),
     deps = [
         "//packages/animations",
+        "//packages/common",
         "//packages/core",
     ],
 )
