+++ packages/animations/browser/test/dsl/style_normalizer/web_animations_style_normalizer_spec.ts
@@ -5,7 +5,7 @@
  * Use of this source code is governed by an MIT-style license that can be
  * found in the LICENSE file at https://angular.io/license
  */
-import {WebAnimationsStyleNormalizer} from '../../../src/dsl/style_normalization/web_animations_style_normalizer';
+import {WebAnimationsStyleNormalizer} from '../../../src/dsl/animation_style_normalizer';
 
 {
   describe('WebAnimationsStyleNormalizer', () => {
