+++ packages/animations/browser/test/render/web_animations/web_animations_driver_spec.ts
@@ -5,9 +5,7 @@
  * Use of this source code is governed by an MIT-style license that can be
  * found in the LICENSE file at https://angular.io/license
  */
-import {browserDetection} from '@angular/platform-browser/testing/src/browser_util';
-
-import {WebAnimationsDriver} from '../../../src/render/web_animations/web_animations_driver';
+import {WebAnimationsDriver} from '../../../src/render/animation_driver';
 import {WebAnimationsPlayer} from '../../../src/render/web_animations/web_animations_player';
 
 {
