+++ packages/core/test/acceptance/renderer_factory_spec.ts
@@ -11,7 +11,7 @@ import {ɵAnimationEngine, ɵNoopAnimationStyleNormalizer} from '@angular/animat
 import {MockAnimationDriver, MockAnimationPlayer} from '@angular/animations/browser/testing';
 import {CommonModule, DOCUMENT} from '@angular/common';
 import {PLATFORM_BROWSER_ID, PLATFORM_SERVER_ID} from '@angular/common/src/platform_id';
-import {Component, DoCheck, NgZone, Renderer2, RendererFactory2, RendererStyleFlags2, RendererType2, ViewEncapsulation} from '@angular/core';
+import {ApplicationRef, Component, DoCheck, NgZone, Renderer2, RendererFactory2, RendererStyleFlags2, RendererType2, ViewEncapsulation} from '@angular/core';
 import {RElement} from '@angular/core/src/render3/interfaces/renderer_dom';
 import {ngDevModeResetPerfCounters} from '@angular/core/src/util/ng_dev_mode';
 import {NoopNgZone} from '@angular/core/src/zone/ng_zone';
@@ -249,8 +249,9 @@ describe('animation renderer factory', () => {
       declarations: [SomeComponentWithAnimation, SomeComponent],
       providers: [{
         provide: RendererFactory2,
-        useFactory: (d: any) => rendererFactory = getAnimationRendererFactory2(d),
-        deps: [DOCUMENT]
+        useFactory: (d: Document, appRef: ApplicationRef) => rendererFactory =
+            getAnimationRendererFactory2(d, appRef),
+        deps: [DOCUMENT, ApplicationRef]
       }]
     });
   });
@@ -319,7 +320,7 @@ describe('animation renderer factory', () => {
   });
 });
 
-function getRendererFactory2(document: any): RendererFactory2 {
+function getRendererFactory2(document: Document): RendererFactory2 {
   const fakeNgZone: NgZone = new NoopNgZone();
   const eventManager = new EventManager([], fakeNgZone);
   const appId = 'app-id';
@@ -335,12 +336,13 @@ function getRendererFactory2(document: any): RendererFactory2 {
   return rendererFactory;
 }
 
-function getAnimationRendererFactory2(document: any): RendererFactory2 {
+function getAnimationRendererFactory2(
+    document: Document, appRef: ApplicationRef): RendererFactory2 {
   const fakeNgZone: NgZone = new NoopNgZone();
   return new ɵAnimationRendererFactory(
       getRendererFactory2(document),
       new ɵAnimationEngine(
-          document.body, new MockAnimationDriver(), new ɵNoopAnimationStyleNormalizer()),
+          document, new MockAnimationDriver(), new ɵNoopAnimationStyleNormalizer(), appRef),
       fakeNgZone);
 }
 
@@ -355,7 +357,7 @@ describe('custom renderer', () => {
   /**
    * Creates a patched renderer factory that creates elements with a shape different than DOM node
    */
-  function createPatchedRendererFactory(document: any) {
+  function createPatchedRendererFactory(document: Document) {
     let rendererFactory = getRendererFactory2(document);
     const origCreateRenderer = rendererFactory.createRenderer;
     rendererFactory.createRenderer = function(element: any, type: RendererType2|null) {
@@ -376,7 +378,7 @@ describe('custom renderer', () => {
       declarations: [SomeComponent],
       providers: [{
         provide: RendererFactory2,
-        useFactory: (document: any) => createPatchedRendererFactory(document),
+        useFactory: (document: Document) => createPatchedRendererFactory(document),
         deps: [DOCUMENT]
       }]
     });
@@ -431,7 +433,7 @@ describe('Renderer2 destruction hooks', () => {
       declarations: [SimpleApp, AppWithComponents, BasicComponent],
       providers: [{
         provide: RendererFactory2,
-        useFactory: (document: any) => getRendererFactory2(document),
+        useFactory: (document: Document) => getRendererFactory2(document),
         deps: [DOCUMENT]
       }]
     });
