+++ packages/platform-browser/animations/src/animations.ts
@@ -12,6 +12,6 @@
  * Entry point for all animation APIs of the animation browser package.
  */
 export {ANIMATION_MODULE_TYPE} from '@angular/core';
-export {BrowserAnimationsModule, BrowserAnimationsModuleConfig, NoopAnimationsModule, provideAnimations, provideNoopAnimations} from './module';
+export {BrowserAnimationsModule, BrowserAnimationsModuleConfig, NoopAnimationsModule, provideAnimations, provideLazyLoadedAnimations, provideNoopAnimations} from './module';
 
 export * from './private_export';
