+++ packages/platform-browser/animations/src/providers.ts
@@ -7,33 +7,12 @@
  */
 
 import {AnimationBuilder} from '@angular/animations';
-import {AnimationDriver, ɵAnimationEngine as AnimationEngine, ɵAnimationStyleNormalizer as AnimationStyleNormalizer, ɵNoopAnimationDriver as NoopAnimationDriver, ɵWebAnimationsDriver as WebAnimationsDriver, ɵWebAnimationsStyleNormalizer as WebAnimationsStyleNormalizer} from '@angular/animations/browser';
-import {DOCUMENT} from '@angular/common';
-import {ANIMATION_MODULE_TYPE, ApplicationRef, Inject, Injectable, NgZone, OnDestroy, Provider, RendererFactory2} from '@angular/core';
+import {AnimationDriver, ɵAnimationEngine as AnimationEngine, ɵAnimationStyleNormalizer as AnimationStyleNormalizer, ɵNoopAnimationDriver as NoopAnimationDriver, ɵNoopAnimationStyleNormalizer as NoopAnimationStyleNormalizer, ɵWebAnimationsDriver as WebAnimationsDriver, ɵWebAnimationsStyleNormalizer as WebAnimationsStyleNormalizer} from '@angular/animations/browser';
+import {ANIMATION_MODULE_TYPE, NgZone, Provider, RendererFactory2} from '@angular/core';
 import {ɵDomRendererFactory2 as DomRendererFactory2} from '@angular/platform-browser';
 
 import {BrowserAnimationBuilder} from './animation_builder';
 import {AnimationRendererFactory} from './animation_renderer';
-
-@Injectable()
-export class InjectableAnimationEngine extends AnimationEngine implements OnDestroy {
-  // The `ApplicationRef` is injected here explicitly to force the dependency ordering.
-  // Since the `ApplicationRef` should be created earlier before the `AnimationEngine`, they
-  // both have `ngOnDestroy` hooks and `flush()` must be called after all views are destroyed.
-  constructor(
-      @Inject(DOCUMENT) doc: any, driver: AnimationDriver, normalizer: AnimationStyleNormalizer,
-      appRef: ApplicationRef) {
-    super(doc.body, driver, normalizer);
-  }
-
-  ngOnDestroy(): void {
-    this.flush();
-  }
-}
-
-export function instantiateDefaultStyleNormalizer() {
-  return new WebAnimationsStyleNormalizer();
-}
 
 export function instantiateRendererFactory(
     renderer: DomRendererFactory2, engine: AnimationEngine, zone: NgZone) {
@@ -41,9 +20,7 @@ export function instantiateRendererFactory(
 }
 
 const SHARED_ANIMATION_PROVIDERS: Provider[] = [
-  {provide: AnimationBuilder, useClass: BrowserAnimationBuilder},
-  {provide: AnimationStyleNormalizer, useFactory: instantiateDefaultStyleNormalizer},
-  {provide: AnimationEngine, useClass: InjectableAnimationEngine}, {
+  {provide: AnimationBuilder, useClass: BrowserAnimationBuilder}, {
     provide: RendererFactory2,
     useFactory: instantiateRendererFactory,
     deps: [DomRendererFactory2, AnimationEngine, NgZone]
@@ -55,6 +32,7 @@ const SHARED_ANIMATION_PROVIDERS: Provider[] = [
  * include them in the BrowserModule.
  */
 export const BROWSER_ANIMATIONS_PROVIDERS: Provider[] = [
+  {provide: AnimationStyleNormalizer, useClass: WebAnimationsStyleNormalizer},
   {provide: AnimationDriver, useFactory: () => new WebAnimationsDriver()},
   {provide: ANIMATION_MODULE_TYPE, useValue: 'BrowserAnimations'}, ...SHARED_ANIMATION_PROVIDERS
 ];
@@ -67,3 +45,24 @@ export const BROWSER_NOOP_ANIMATIONS_PROVIDERS: Provider[] = [
   {provide: AnimationDriver, useClass: NoopAnimationDriver},
   {provide: ANIMATION_MODULE_TYPE, useValue: 'NoopAnimations'}, ...SHARED_ANIMATION_PROVIDERS
 ];
+
+/**
+ * Same as BROWSER_ANIMATIONS_PROVIDERS but with lazy loaded modules and without AnimationBuilder
+ *
+ */
+export const LAZY_LOADED_ANIMATIONS_PROVIDERS: Provider[] = [
+  {
+    provide: RendererFactory2,
+    useFactory: instantiateRendererFactory,
+    deps: [DomRendererFactory2, AnimationEngine, NgZone]
+  },
+  {provide: ANIMATION_MODULE_TYPE, useValue: 'BrowserAnimations'}
+];
+
+/**
+ * Provider for the Browser AnimationBuilder
+ */
+export const BROWER_ANIMATION_BUILDER_PROVIDER: Provider = {
+  provide: AnimationBuilder,
+  useClass: BrowserAnimationBuilder
+};
